requirements.txt
PySide6>=6.6.0
onnx>=1.15.0
onnxruntime>=1.17.0
Pillow>=10.0.0
numpy>=1.23
